feat: move boundary line functions to line-helpers.ts and update exports
